added action and reducer for example
