+++ src/renderer/actions/like.js
@@ -1,0 +1,5 @@
+import { createAction } from 'redux-actions';
+
+export default {
+  likeIt: createAction('LIKE_IT'),
+};
