--- src/renderer/actions/user.js
@@ -1,6 +1,0 @@
-import { createAction } from 'redux-actions';
-
-export default {
-  login: createAction('USER_LOGIN'),
-  logout: createAction('USER_LOGOUT'),
-};
